ppt-pipeline/requirements.txt
flask
python-pptx
python-dotenv
pdfplumber
openai
moviepy
Pillow
pdf2image
kokoro
imageio-ffmpeg
numpy
soundfile
gtts
